% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_individual_sections_on_the_slide.R
\name{assign_ID_based_on_overlap}
\alias{assign_ID_based_on_overlap}
\title{Assign a spot ID based on whether it overlaps with a selection of polygons}
\usage{
assign_ID_based_on_overlap(
  sce.obj,
  subset_by = "Sample",
  coord_list,
  id_name = "sec",
  plot_overlap = FALSE
)
}
\arguments{
\item{subset_by}{Default: "Sample"}

\item{coord_list}{Must contain one element per entry of "subset_by"}

\item{id_name}{Default: "sec"}

\item{plot_overlap}{Boolean. Default: FALSE. If set to TRUE, the function
will not change the sce.obj, but will simply plot the fit for visual inspection}

\item{sce.object}{SingleCellExperiment Object}
}
\value{
SCE object with additional column in the colData that indicates the
name of the polygon that a given spot overlapped with
}
\description{
Assign a spot ID based on whether it overlaps with a selection of polygons
}
\examples{
\dontrun{
sec.coords <-list()
smp <- "pN3A"
sce.norm[, sce.norm$Sample == smp] \%>\% reducedDim(., "spots") \%>\% plot
sec.coords[[smp]]$secI <- locator(15)
sec.coords[[smp]]$secII <- locator(15)
}

}
\seealso{
Wrapper around \code{\link{determine_overlap}},
 \code{\link{create_polygons_from_xy}}
}
