% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markvario.R
\name{ComputeRMetric}
\alias{ComputeRMetric}
\title{Computes the metric at a given r (radius) value}
\usage{
ComputeRMetric(mv, r.metric = 5)
}
\arguments{
\item{mv}{Results of running markvario}

\item{r.metric}{r value at which to report the "trans" value of the mark
variogram}
}
\value{
Returns a data.frame with r.metric values
}
\description{
Computes the metric at a given r (radius) value
}
\details{
Taken as is from Seurat 3.1
}
