% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_individual_sections_on_the_slide.R
\name{create_polygons_from_xy}
\alias{create_polygons_from_xy}
\title{Create polygon objects based on x and y coordinates}
\usage{
create_polygons_from_xy(coord_list)
}
\arguments{
\item{coord_list}{list of x and y coordinates per region of interest, where each
element of the list coorresponds to a region, containing x and y coordinates
as another list or data.frame. It should be a named list.}
}
\value{
SpatialPolygons object
}
\description{
Create polygon objects based on x and y coordinates
}
