% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_individual_sections_on_the_slide.R
\name{determine_overlap}
\alias{determine_overlap}
\title{Determine the overlap between spots from an SCE and user-defined polygons}
\usage{
determine_overlap(sce.obj, coord_list, plot_overlap = FALSE)
}
\arguments{
\item{sce.obj}{SCE object containing the cells that should be assigned an
ID based on which polygon defined in coord_list they overlap with}

\item{coord_list}{list of x and y coordinates per region of interest, where each
element of the list corresponds to a region, containing x and y coordinates
as another list or data.frame. It should be a named list.}

\item{plot_overlap}{Boolean. Default: FALSE. If set to TRUE, the function
will not change the sce.obj, but will simply plot the fit for visual inspection}
}
\value{
if plot_overlap = FALSE, this will return an named vector of
points (cells) overlapping with the polygons
}
\description{
Determine the overlap between spots from an SCE and user-defined polygons
}
