% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_spots.R
\name{fix_path}
\alias{fix_path}
\title{Helper function for correcting file paths}
\usage{
fix_path(path)
}
\value{
Path with "/" at the end (string)
}
\description{
Ensures that every file path string has a "/" at the end.
}
