% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spots_prepareDRP.R
\name{get_SpotPlotDRP}
\alias{get_SpotPlotDRP}
\title{Prepare the "DimRedPlot" object for plotting spot information}
\usage{
get_SpotPlotDRP(
  object,
  which_reddim = "spots",
  X = NULL,
  Y = NULL,
  image_data = NULL,
  color_by = NULL,
  shape_by = NULL,
  size_by = NULL,
  label_by = NULL,
  circle_by = NULL,
  exprs_values = "counts",
  add_cell_info = NULL
)
}
\arguments{
\item{object}{\code{SingleCellExperiment} object}

\item{which_reddim}{accessor to retrieve the list of choice for the reduced
dimension coordinates from \code{reducedDims(object)}. See \code{details}.}

\item{X}{directly supply the X coordinates, e.g. \code{pca_res$PC1}; if this
is used, \code{which_reddim} is ignored.
\code{X} should be of the same length (and order!) as \code{colData(object)}
(or \code{pData2.0(object)}).}

\item{Y}{directly supply the Y coordinates, e.g. \code{pca_res$PC2}.
Should be of the same length (and order!) as \code{colData(object)}.
\code{which_reddim} and \code{which_pcs} will be ignored if \code{X} and
\code{Y} are specified.}

\item{image_data}{list with the grob, height and width of the PNG images
of the corresponding tissue (usually the low resolution version). It should
also contain a column named "slide", which should indicate a unique name per
slide. This column will be used to facet the final plot. Make sure that its
values are actually part of the sce.object's colData, too.
See \code{\link{reading_image_with_mods}} for details on how to generate an
appropriate list.}

\item{color_by}{entry of either \code{colData(object)}, \code{rowData(object)},
\code{rownames(object)}, which will be used to assign either discrete or
continous color schemes. Typically, only one value is accepted here, except for
gene names, where, if multiple genes are indicated, the median expression across
all genes will be shown per cell. Alternatively, you can supply a \code{list}
with a label stored in \code{list$title} and values stored in \code{list$result}.
(\code{result} should be as long as \code{dim(object)[2]}. Default: \code{NULL}}

\item{shape_by}{default: \code{NULL}}

\item{size_by}{default: \code{NULL}}

\item{label_by}{default: \code{NULL}}

\item{circle_by}{Used for encircling points of interest; default: \code{NULL}.
If \code{color_by} is a list, this will be ignored here because the list will
be used to subset the final data.frame once the actual plot is being constructed.}

\item{exprs_values}{types of expression values that will be used for plotting if a gene name is being
specified in either \code{color_by}, \code{shape_by}, \code{size_by}}

\item{add_cell_info}{if you want to add columns for certain entries of
\code{colData(object)}, specify their names here. Default: NULL}
}
\value{
DimRedPlot object: a simple S3-based object, i.e. a list with:
\enumerate{
\item \code{plot_data}: data.frame fit for ggplot2-based plotting
\item \code{label_data}: additional stuff that may be used for labelling the
x and y axes of the final plot
\item \code{factors}: names of the plotting variables, e.g. which values
were used for defining the coloring etc.
}
}
\description{
Wrapper function for extracting the X and Y coordinates for the
plot as well as the variables for coloring, shape assignment etc.
This function expects a SingleCellExperiment object.
}
\details{
\emph{which_reddim} The reducedDims slot of the SingleCellExperiment object
is a list (\code{\link{SimpleList}}. To know which coordinates to extract,
the accessor for the specific reducedDims list member must be given.
}
\seealso{
\code{\link{plt.SpotPlot}} for plotting the resulting DRP and
\code{\link{plot_Spot_from_sce}} for a wrapper to generate the DRP and the
plot in one go starting from an SCE and image_data
}
