% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markvario.R
\name{make_pp}
\alias{make_pp}
\title{Generate a point pattern object}
\usage{
make_pp(sce.obj, coords_accessor)
}
\arguments{
\item{sce.obj}{SCE object with XY coordinates in reducedDim slot and normalized expression values
that can be used for testing whether individual genes show expression patterns that correlate with
the spatial location of the respective cells}

\item{coords_accessor}{Indicate which reducedDim slot should be accessed to obtain a matrix of x
and y coordinates of the cells}
}
\description{
Generate a point pattern object
}
