% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spots.R
\name{plt.SpotPlot}
\alias{plt.SpotPlot}
\title{Visualizing tissue coordinates}
\usage{
plt.SpotPlot(
  drp_object,
  X = "x_axs",
  Y = "y_axs",
  color_by = NULL,
  color_by_factor = FALSE,
  shape_by = NULL,
  size_by = NULL,
  label_by = NULL,
  label_size = 4,
  circle_by = NULL,
  which_quantile = 1,
  ignore_drp_labels = FALSE,
  theme_size = 14,
  legend = "auto",
  alpha = 0.75,
  remove_rug = FALSE,
  set_colors = TRUE,
  set_fill_colors = TRUE
)
}
\arguments{
\item{drp_object}{a \code{\link{DimRedPlot}} object}

\item{X}{name of the drp_object column that contains the values for the x axis}

\item{Y}{name of the drp_object column that contains the values for the y axis}

\item{color_by}{specify the column of \code{drp_object$plot_data} whose
entries will be used to assign either discrete or continous color schemes.}

\item{color_by_factor}{if set to TRUE, the values supplied to \code{color_by}
will be interpreted as factors (as long as there are fewer than 14 individual
values). Default: FALSE}

\item{shape_by}{the factor to assign the shape of the points to; max. 10 levels
are accepted for \code{shape_by}. If you want to specify the (same) shape for
\emph{all} points, use an integer (default shape is 19).}

\item{size_by}{the factor to assign the size of the points to;
default: \code{NULL}. If you want to specify the (same) size for \emph{all}
points, use one integer (the default size is 3).}

\item{label_by}{factor (part of colData(sce)) used to label groups of cells.
A typical example would be "cluster" or "condition". Default: NULL.}

\item{label_size}{default: 4}

\item{circle_by}{specify the column of \code{drp_object$plot_data} whose
entires will be interpreted as factors and used to draw circles
around the points belonging to the same factor instance (e.g. "condition").
Alternatively, supply a list (!) where \code{list$cells} indicates the cell names
(= rownames of \code{drp_object$plot_data}) around which a circle (without fill)
will be drawn. Example: \code{list(cells = head(colnames(tiny_sce))}}

\item{which_quantile}{default: 1}

\item{ignore_drp_labels}{If set to TRUE, \emph{all} entries of
\code{drp_object$factors} will be ignored. If you want to ignore only selected
labels, indicate their name (e.g. c("color_by", "exprs_val_type")).
The default setting (\code{FALSE}) will make use of the information from
the drp_object \emph{unless} specified here via \code{color_by}, \code{shape_by},
\code{size_by}, \code{circle_by}.}

\item{theme_size}{font size for the plot}

\item{legend}{choose wether there should be no legend ("none"), a legend for
every factor or whether the legend will only show up for those factors that
have more than 1 level ("auto", default setting).}

\item{alpha}{define the opacity of the points, default: .65}

\item{remove_rug}{Define whether the bars at both axes that
represent the two 1d marginal distributions should be removed. Default: FALSE.}

\item{set_colors}{Set to \code{FALSE} if you want to add your own color scheme
to the resulting plot. The default behavior is to try to automatically assign
an optimized coloring scheme. Default: \code{TRUE}.}

\item{set_fill_colors}{Set to \code{FALSE} if you want to add your own color
scheme to the resulting plot \emph{for the background circles}. The default
behavior is to try to automatically assign an optimized coloring scheme.}
}
\value{
\code{ggplot2} object (a plot)
}
\description{
This will generate a ggplot starting from a \code{DimRedPlot} object.
See \code{\link{get_SpotPlot.sce}} to start from a \code{SingleCellExperiment}
object.
}
\seealso{
\code{\link{get_SpotPlot.sce}}, which generates the DimRedPlot
object that will be used here.
}
