% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_spots.R
\name{reading_image}
\alias{reading_image}
\title{Loading PNG information for plotting with ggplot2}
\usage{
reading_image(
  image_path,
  which_res = c("lowres", "hires"),
  slide_name = "Sample.X"
)
}
\arguments{
\item{image_path}{path to the image to be loaded and modified}

\item{which_res}{either one of "lowres" or "hires"}

\item{slide_name}{Default: "Sample.X"}
}
\value{
A tibble with the sample name and the PNG information (grob)
}
\description{
This function loads the PNG image of the Visium slide and returns
a tibble that contains the corresponding information that can be used to plot
the image within the ggplot2 framework (see \code{\link{geom_visium}}).
}
\examples{
\dontrun{
img1 <- reading_image("data/pN3A/spatial/", slide_name = "pN3A",which_res = "lowres")
bc_info <- read_spot_coordinates("data/pN3A/spatial/", slide_name = "pN3A", rm_empty_spots = FALSE)

}

}
\seealso{
\code{\link{geom_visium}}
}
