% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_spots.R
\name{reading_image_with_mods}
\alias{reading_image_with_mods}
\title{Modify the image to improve plotting}
\usage{
reading_image_with_mods(
  image_path,
  slide_name = "Sample.X",
  which_res = c("lowres", "hires"),
  testing_mods = TRUE,
  brightness = 450,
  saturation = 120,
  hue = 90,
  ...
)
}
\arguments{
\item{image_path}{path to the image to be loaded and modified}

\item{slide_name}{Default: "Sample.X"}

\item{which_res}{either one of "lowres" or "hires"}

\item{testing_mods}{Default: TRUE. To determine the best settings, use
the function with testing_mods = TRUE, which will simply print the results of
the modification. Once you're satisfied with the outcome, set testing_mods =
FALSE}

\item{brightness}{modulation of brightness as percentage of the current value
(100 for no change)}

\item{saturation}{modulation of saturation as percentage of the current value
(100 for no change)}

\item{hue}{modulation of hue is an absolute rotation of -180 degrees to +180
degrees from the current position corresponding to an argument range of 0 to
200 (100 for no change)}

\item{...}{additional parameters for \code{\link{magick::image_modulate}}}
}
\value{
If testing_mods is set to FALSE, then the function will return a
tibble that includes the modified glob of the original image.
}
\description{
This function allows you to read in an image and change key parameters
such as brightness, saturation and hue via \code{\link{magick::image_modulate}}.
}
\details{
Requires the magick R package. To determine the best settings, use
the function with testing_mods = TRUE, which will simply print the results of
the modification. Once you're satisfied with the outcome, set testing_mods =
FALSE, which will result in a tibble that's equivalent to the one generated
by \code{\link{reading_image}}.
}
\seealso{
\code{\link{magick::image_modulate}}, \code{\link{reading_image}}
}
