% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markvario.R
\name{run_markvario}
\alias{run_markvario}
\title{Run the mark variogram computation}
\usage{
run_markvario(
  sce.obj,
  coords_accessor = "spots",
  exprs_values = "log1p_sctransform",
  genes = NULL,
  r.metric = 5,
  return_SCE = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sce.obj}{SCE object with XY coordinates in reducedDim slot and normalized expression values
that can be used for testing whether individual genes show expression patterns that correlate with
the spatial location of the respective cells}

\item{coords_accessor}{Indicate which reducedDim slot should be accessed to obtain a matrix of x and y coordinates
of the cells}

\item{exprs_values}{Which assay should be used from the SCE object. Default: "log1p_sctransform"}

\item{genes}{if you want to select a subset of genes for the analysis, specify their names here. Default: NULL will
use all genes in the SCE object.}

\item{r.metric}{r value at which to report the "trans" value of the mark variogram. Default: 5}

\item{verbose}{Default: TRUE}

\item{...}{Additional Arguments passed to \code{\link{spatstat::markvario}}}
}
\value{
The SCE object will be returned with additional entries in the rowData:
 "r.metric.5" and "spatially.variable_mv.rank"
}
\description{
Wraps the functionality of \code{\link{spatstat::markvario}}.
Inspired by [Trendsceek (Endsgard, Johnsson, Sandberg, 2018)](https://www.nature.com/articles/nmeth.4634),
which models spatial transcriptomics data as a mark point process and computes a ‘variogram’, which
 identifies genes whose expression level is dependent on their spatial location.
 The calculation of the r-metric was copied from Seurat v. 3.1 (\code{ComputeRMetric}).
}
\details{
This process calculates gamma(r) values measuring the dependence between
 two spots that are a certain "r" distance apart.
Spatial data can be made up of point patterns (location) with attached values (e.g. expression).
The mark variogram indicates how similar two marks (here: genes) within a certain distance from each other are.
}
\references{
Daniel Edsgärd, Per Johnsson & Rickard Sandberg.
Identification of spatial expression trends in single-cell gene expression data
\emph{Nature Methods}, 15, 339–342 (2018).
doi: 10.1038/nmeth.4634

Adrian Baddeley, Ege Rubak and Rolf Turner:
Spatial Point Patterns -- Methodology and Applications with R.
CRC Press, 2015. ISBN 978-1-4822-1020-0. 810 pp.
\url{http://www.spatstat.org/}
}
